+++ new-ui/src/js/components/Search.jsx
@@ -8,11 +8,24 @@ export default class Search extends React.Component {
 
         this.state = {
           bookings : [],
-          hotels : []
-        }
+		  hotels : [],
+		  query : ""
+		}
+		
+		this.doSearch = this.doSearch.bind(this);
     }
 
     componentDidMount() {
+		this.doSearch()
+	}
+	
+	componentDidUpdate() {
+		if(this.state.query != this.props.location.search){
+			this.doSearch();
+		}
+	}
+
+	doSearch() {
 		fetch('http://localhost:3002/search' + this.props.location.search, {
 			headers: {
 				'Accept': 'application/json',
@@ -21,9 +34,21 @@ export default class Search extends React.Component {
 		})
 		.then(res => res.json())
 		.then(res => {
-			this.setState(res)
-		})
-    }
+			if(res.bookings){
+				this.setState({bookings : res.bookings})
+			} else {
+				this.setState({bookings : []})
+			}
+
+			if(res.hotels){
+				this.setState({hotels : res.hotels})
+			} else {
+				this.setState({hotels : []})
+			}
+			
+			this.setState({query : this.props.location.search})
+		});
+	}
 
     render() {
     	return(
@@ -40,7 +65,7 @@ export default class Search extends React.Component {
                 </div>
 				{this.state.hotels.map((hotel, index) => {
 					return <div className="row detail" key={"hotel-search-" + index}>
-								<div className="col-sm-12 searchResult"><Link to={"/hotel/" + hotel.hotelid }>{hotel.name}</Link></div>
+								<Link to={"/hotel/" + hotel.hotelid }><div className="col-sm-12 searchResult" style={{paddingBottom : 10 + 'px'}}>{hotel.name}</div></Link>
 							</div>
 				})}
 				<div className="row">
@@ -50,9 +75,11 @@ export default class Search extends React.Component {
                 </div>
 				{this.state.bookings.map((booking, index) => {
 					return <div className="row detail searchResult" key={"booking-search-" + index}>
-								<div className="col-sm-8"><p>{booking.firstname} {booking.lastname}</p></div>
-								<div className="col-sm-2" style={{textAlign: 'center'}}><p>{booking.bookingdates.checkin.split('T')[0]}</p></div>
-								<div className="col-sm-2" style={{textAlign: 'center'}}><p>{booking.bookingdates.checkout.split('T')[0]}</p></div>
+								<Link to={"/hotel/" + booking.hotelid}>
+									<div className="col-sm-8"><p>{booking.firstname} {booking.lastname}</p></div>
+									<div className="col-sm-2" style={{textAlign: 'center'}}><p>{booking.bookingdates.checkin.split('T')[0]}</p></div>
+									<div className="col-sm-2" style={{textAlign: 'center'}}><p>{booking.bookingdates.checkout.split('T')[0]}</p></div>
+								</Link>
 							</div>
 				})}
 			</div>
